BulletinBoardPage.tsxを作成
