+++ frontend/yui-loop/src/types/WaitingPassengerType.ts
@@ -1,3 +1,8 @@
 export interface WaitingPassengerType {
-    message: string, 
+    month: string
+    date: string
+    time: string
+    start: string
+    destination: string
+    capacity: string
 }
